refactor(scan,scanner): move scan and scanner controllers to api pkg
