+++ src/api/scan/all_handler.go
