+++ src/api/scan/controller.go
