+++ src/api/scanner/base_controller.go
