+++ src/core/api/scan_all.go
@@ -6,14 +6,14 @@ import (
 	"strconv"
 	"strings"
 
+	"github.com/goharbor/harbor/src/api/scan"
+	"github.com/goharbor/harbor/src/api/scanner"
 	common_job "github.com/goharbor/harbor/src/common/job"
 	cm "github.com/goharbor/harbor/src/common/models"
 	"github.com/goharbor/harbor/src/core/api/models"
 	"github.com/goharbor/harbor/src/jobservice/job"
 	"github.com/goharbor/harbor/src/pkg/q"
 	"github.com/goharbor/harbor/src/pkg/scan/all"
-	"github.com/goharbor/harbor/src/pkg/scan/api/scan"
-	"github.com/goharbor/harbor/src/pkg/scan/api/scanner"
 	"github.com/pkg/errors"
 )
