+++ src/core/api/scanners.go
@@ -18,8 +18,8 @@ import (
 	"fmt"
 	"net/http"
 
+	s "github.com/goharbor/harbor/src/api/scanner"
 	"github.com/goharbor/harbor/src/pkg/q"
-	s "github.com/goharbor/harbor/src/pkg/scan/api/scanner"
 	"github.com/goharbor/harbor/src/pkg/scan/dao/scanner"
 	"github.com/pkg/errors"
 )
