+++ src/core/middlewares/vulnerable/handler.go
@@ -17,15 +17,16 @@ package vulnerable
 import (
 	"net/http"
 
+	"net/http/httptest"
+
+	sc "github.com/goharbor/harbor/src/api/scan"
 	"github.com/goharbor/harbor/src/common/models"
 	"github.com/goharbor/harbor/src/common/utils/log"
 	"github.com/goharbor/harbor/src/core/middlewares/util"
-	sc "github.com/goharbor/harbor/src/pkg/scan/api/scan"
 	"github.com/goharbor/harbor/src/pkg/scan/report"
 	v1 "github.com/goharbor/harbor/src/pkg/scan/rest/v1"
 	"github.com/goharbor/harbor/src/pkg/scan/vuln"
 	"github.com/pkg/errors"
-	"net/http/httptest"
 )
 
 type vulnerableHandler struct {
